🧪 test(core): own canonical evaluator fixtures
