+++ .prettierignore
@@ -10,4 +10,4 @@ yarn.lock
 test-app/dist
 test-app/node_modules
 test-app/*.tgz
-packages/node-server/test/ffe-system-test-data
+packages/core/test/ffe-system-test-data
