+++ packages/core/test/TestCaseResult.types.ts
@@ -1,0 +1,17 @@
+import type { FlagValue } from '@openfeature/core'
+import type { PrecomputedFlagMetadata } from '../src/configuration'
+import type { VariantType } from '../src/evaluation'
+
+export interface TestCase {
+  flag: string
+  variationType: VariantType
+  defaultValue: FlagValue
+  targetingKey: string | null
+  attributes: Record<string, unknown>
+  result: {
+    value: FlagValue
+    reason: string
+    variant?: string
+    flagMetadata?: PrecomputedFlagMetadata
+  }
+}
