--- packages/node-server/test/TestCaseResult.types.ts
@@ -1,17 +1,0 @@
-import type { PrecomputedFlagMetadata } from '@datadog/flagging-core'
-import type { EvaluationContextValue, FlagValue } from '@openfeature/server-sdk'
-import type { VariantType } from '../src/configuration/ufc-v1'
-
-export interface TestCase {
-  flag: string
-  variationType: VariantType
-  defaultValue: FlagValue
-  targetingKey: string | null
-  attributes: Record<string, EvaluationContextValue>
-  result: {
-    value: FlagValue
-    reason: string
-    variant?: string
-    flagMetadata?: PrecomputedFlagMetadata
-  }
-}
