--- packages/node-server/test/flags-v1.spec.ts
@@ -1,123 +1,0 @@
-import type { Channel } from 'node:diagnostics_channel'
-import fs from 'node:fs'
-import path from 'node:path'
-import type { ExposureEvent } from '@datadog/flagging-core'
-import type { EvaluationContext, EvaluationDetails, FlagValue, JsonValue, Logger } from '@openfeature/core'
-import { OpenFeature } from '@openfeature/server-sdk'
-import type { UniversalFlagConfigurationV1 } from '../src/configuration/ufc-v1'
-import { DatadogNodeServerProvider } from '../src/provider'
-import type { TestCase } from './TestCaseResult.types'
-
-type ExposureChannelListener = (message: ExposureEvent, name: string | symbol) => void
-
-describe('Universal Flag Configuration V1', () => {
-  let logger: Logger
-
-  const exposureChannelMessageHandler: jest.Mock<ExposureChannelListener> = jest.fn()
-
-  const exposureChannel = {
-    hasSubscribers: true,
-    publish: jest.fn((message: ExposureEvent) => {
-      exposureChannelMessageHandler(message, 'ffe:exposure:submit')
-    }),
-    subscribe: jest.fn(),
-  } as unknown as Channel<ExposureEvent, ExposureEvent>
-
-  beforeEach(() => {
-    logger = {
-      error: console.error,
-      warn: console.warn,
-      info: console.info,
-      debug: jest.fn(),
-    }
-    exposureChannelMessageHandler.mockClear()
-  })
-
-  const getUFC = (): UniversalFlagConfigurationV1 => {
-    const ufcJson = fs.readFileSync(path.join(__dirname, './ffe-system-test-data', 'ufc-config.json'), 'utf8')
-    return JSON.parse(ufcJson) as UniversalFlagConfigurationV1
-  }
-
-  const getTestCaseFileNames = (): string[] => {
-    const fixtureFiles = fs.readdirSync(path.join(__dirname, './ffe-system-test-data/evaluation-cases')).sort()
-    if (fixtureFiles.length === 0) {
-      throw new Error('FFE fixture submodule is missing or empty')
-    }
-    return fixtureFiles
-  }
-
-  const getTestCases = (testCaseFileName: string): TestCase[] => {
-    const testCases = fs.readFileSync(
-      path.join(__dirname, './ffe-system-test-data/evaluation-cases', testCaseFileName),
-      'utf8'
-    )
-    return JSON.parse(testCases) as TestCase[]
-  }
-
-  const evaluateDetails = async (
-    testCase: TestCase,
-    context: EvaluationContext
-  ): Promise<EvaluationDetails<boolean | string | number | JsonValue>> => {
-    const ufc = getUFC()
-    const provider = new DatadogNodeServerProvider({
-      exposureChannel: exposureChannel,
-    })
-    provider.setConfiguration(ufc)
-    OpenFeature.setProvider(provider)
-    OpenFeature.setLogger(logger)
-    OpenFeature.setContext(context)
-    const client = OpenFeature.getClient()
-    if (testCase.variationType === 'BOOLEAN') {
-      return await client.getBooleanDetails(testCase.flag, testCase.defaultValue as boolean)
-    }
-    if (testCase.variationType === 'STRING') {
-      return await client.getStringDetails(testCase.flag, testCase.defaultValue as string)
-    }
-    if (testCase.variationType === 'INTEGER' || testCase.variationType === 'NUMERIC') {
-      return await client.getNumberDetails(testCase.flag, testCase.defaultValue as number)
-    }
-    if (testCase.variationType === 'JSON') {
-      return await client.getObjectDetails(testCase.flag, testCase.defaultValue as Record<string, FlagValue>)
-    }
-    throw new Error(`Unsupported variation type: ${testCase.variationType}`)
-  }
-
-  describe.each(getTestCaseFileNames())('should evaluate for %s', (testCaseFileName) => {
-    const testCases = getTestCases(testCaseFileName)
-    const testCasesWithContext = testCases.map((testCase) => ({
-      contextString: JSON.stringify({
-        targetingKey: testCase.targetingKey,
-        ...testCase.attributes,
-      }),
-      testCase,
-    }))
-
-    it.each(testCasesWithContext)('with context $contextString', async ({ contextString, testCase }) => {
-      const context = JSON.parse(contextString)
-      const details = await evaluateDetails(testCase, context)
-      expect(details.value).toEqual(testCase.result.value)
-      expect(details.reason).toEqual(testCase.result.reason)
-      if (testCase.result.flagMetadata?.doLog) {
-        expect(exposureChannelMessageHandler).toHaveBeenCalledWith(
-          {
-            timestamp: expect.any(Number),
-            allocation: {
-              key: testCase.result.flagMetadata?.allocationKey,
-            },
-            flag: {
-              key: testCase.flag,
-            },
-            variant: {
-              key: testCase.result.variant,
-            },
-            subject: {
-              id: testCase.targetingKey,
-              attributes: testCase.attributes,
-            },
-          },
-          'ffe:exposure:submit'
-        )
-      }
-    })
-  })
-})
